Move generateAutoValueMappers into a new class
